requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.9
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
